connessione con il database fatta... provatelo per vedere se funziona basta che guardate la schermata del log cat mentre eseguite l'app ho messo dei log.i che vi informano cosa sta facendo

+++ NewFolderForUpload/iWish/src/_iWish/MyUtente.java
@@ -15,29 +15,28 @@ public interface MyUtente {
 	String getCity();
 	String getSex();
 	String getTypeUser();
-	double getHeight();
-	double getWeight();
+	int getHeight();
+	int getWeight();
 	String getEmail();
 	String getPassword();
 	String getQuestion();
 	String getAnswer();
 	Context getC();
-	List<Double> getListWeight();
+	List<Integer> getListWeight();
 	void setKeyUtente(long keyUtente);
 	void setC(Context c);
 	void setName(String name);
 	void setSurname(String surname);
-
 	void setCity(String city);
 	void setEmail(String  email);
 	void setSex(String sex);
-	void setHeight(double height);
-	void setWeight(double weight);
+	void setHeight(int height);
+	void setWeight(int weight);
 	void setTypeUser(String typeUser);
 	void setPassword(String  password);
 	void setQuestion(String question);
 	void setAnswer(String answer);
-	void setListWeight(Double weight);
+	void setListWeight(int weight);
 	void setBirthday(String birthday);
 	
 }
